+++ SlyceSDKDemo/SlyceSDKDemo/CustomHistoryViewController.h
@@ -1,0 +1,13 @@
+//
+//  CustomHistoryViewController.h
+//  SlyceSDKDemo
+//
+//  Created by HADI MRIH on 11/01/2016.
+//  Copyright © 2016 BuyCode. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface CustomHistoryViewController : UIViewController
+
+@end