+++ release_notes.md
@@ -1,6 +1,18 @@
 Slyce SDK Changelog
 ===================================
 
+
+
+2.4 - 2016/01/17
+------------------
+
+* ADDED: shouldUseContinuousRecognition2D property to enable/disable continuous recognition for 2D products functionality.
+* ADDED: shouldUseContinuousRecognitionBarcodes property to enable/disable continuous recognition for barcodes functionality.
+* ADDED: shouldPauseScannerDelayTime property to override the default '3 sec' delay time for resuming the auto scanner automatincly after a successful 1D/2D detection.
+* ADDED: setCustomHelpViewController used to override the default 'help' viewController with your custom viewController.
+* ADDED: setCustomNotFoundViewController used to override the default 'not found' viewController with your custom viewController.
+* ADDED: setCustomBtnWithVC use this method to add your own custom button with a custom viewController to the SFCameraViewController(Full UI mode).
+
 2.3.3 - 2016/01/08
 ------------------
 * FIXED: Minor bug fixes